requirements.txt
python-amazon-paapi>=1.1.0
keepa>=1.3.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0
typer>=0.9.0
apscheduler>=3.10.0
httpx>=0.25.0
